refactor: extract 'LastDeckIdRepository'
